playNoteColumn refactor v1
main focus of refactor was to improve readibility + incorporate the
change where the only chord effects are v and ^ to signify down or up
strums respectively.

We also changed how far apart notes are in a strum based on the bpm.

Additional notes about how to structure playNote() to be as modular
as possible based on the effects of the note.