playNote refactor V1
Migrated to a much more modular approach to applying effects to a note.
We now are creating separate functions for each effect, and then applying
and/or chaining them togther before sending the final result to the
audioContext destination (speakers).

Also fixed setTimeout duration on tethered effects, was doing (bpm/60)
instead of (60/bpm) which resulted in the opposite effect that we wanted.