Found that I can access "detune" param
Refactor that starts to convert effect functions to take advantage of
modifying the detune param of an AudioBufferSourceNode. Also takes a more
reasonable approach to handling "tethered" effects where we only
apply it to the start of the "tethered" or second note in the pair, and
then shift into regular pitch of the second note.

The next commit should migrate the rest of the detune related effects