Exported state from useSound migration
This was a silly mistake on my part, I didn't realize that in the ~3
components that use the useSound hook, each one was truly separate
from each other, when I was exporting state thinking that they would be
consistent across all components. This commit aims to fix that by moving
the state into the TabStore, which is a global store that is consistent
across all components.