Moved state from useSound -> TabStore
I had a lapse in judgement that I only realized now, where despite the
hook form factor of useSound, since it had to be declared in multiple
components, they would all be disjointed in their state.

This commit moves the state to the TabStore, which guarentees that
all components which use "useSound()" will be in sync.