Most effects revamped
Primary change was for tethered effects (hpv\), which can now be
stretched out to (3 h 5) instead of limited just to the next immediate
note (3h 5). Also added ability to "arbitrarily" slide into/out from
a note (/3 and 3/), which are treated as bends within the engine.

Also improved most effects to be more accurate, although it is still a
work in progress!