Slider + Header UI/UX improvements
This commit primarily focuses on getting the audio controls/header to be
out of the user's way on mobile devices since space is at a premium.

Alongside the complex handling of exactly when to autohide/show the
controls/header, the audio controls component got some nice ui tweaks.

More specifically, whenever there is a route change any playing audio
will be stopped and if appropriate the audio controls will be hidden.