Upgraded bend effects + added release effect
Before we only had a static bend effect, which increased the pitch of
the note by a whole step, now we can specify the bend amount by passing
a target fret to bend to in the next column over. Releasing from a bend
was also added, which is just returning the bent note to its original
pitch.

We also made some architectural changes to the metadata surrounding
tethered effects since it was too restrictive to convey what we needed
with the bends.