Sound generation/playback refactor
We took a very roundabout way of handling audio generation and playback
early on due to our lack of fundamental understanding of what we needed
and how unfamiliar we were with how Zustand could help us centralize our
state + functions in a reasonable way.

Our initial approach involved creating a custom hook that housed all of
the logic for generating and playing and pausing audio. But to do this
it needed quite a lot of duplicate state in the form of refs at the top
of the hook, since react state updates wouldn't be accessible/reflected
for blocks of code within an async for loop (generated audio).

This approach takes the core functions from the useSound() hook and puts
them directly in the tabStore, and utilizes zustand's get() method to
access the current state of the store with much less hassle. Ideally we
need to split the store up into different slices since it is a
behemouth, but for now this is a good step in the right direction.

As a side note, generated sound functions/effects/helpers were abstracted
into their own file while cut down on clutter and I removed imported
zustand state from the components that previously needed it to pass into
the generated sound functions.