Infinite fetching optimizations
Implemented a much better ux + ui for infinite fetching loading state.
In general we got rid of the spinner and will now just extend the skeleton
version of the elements while fetching more data. We also improved
the optimistic updating mechanism while liking/unliking a tab to sort
the data instantly (although there are still some edge cases to rough out).

We now only fetch the bare minimun data to display the tab, and will
on demand fetch the tabData to play the song if necessary. This should
improve the performance on slower connections.