Taking a different screenshot approach
Every variation of puppeteer w/ next.js api routes simply was too large
for vercel/aws lambda limits (50MB), so we are using html2canvas with a
couple of extra tricks to basically get the same result.

Utilizes a portal for the <TabPreview /> to be at the bottom of <Tab />
so that it isn't rendered in user's viewport causing jarring ux.