Maximizing vertical space on tab section previews
This is kind of a messy commit, but we had way too much _consistent_
empty space in our tab section previews, so this commit now has all
appropriate jsx and logic to only use the minimal space necessary to
render the rows. This should help out on mobile especially since it was
hard to figure out what you might be playing next when it was totally
offscreen. Ideally we would shrink the font-size down even further so
that we could further this idea, but trying this for now.