Add a chord (relative) + adj note length per chord
This commit focuses on quality of life improvements for editing tab
sections. Before our only option to add another chord was to add 8 at a
time with the "Extend tab" button, however it was tedious to add 8, move
the one you wanted to the right place, and then delete the rest.

Next up, we aren't fully chained to the per measure line bpm adjustments
anymore. What is often required to achieve a dynamic sound is to have
certain notes played for longer or shorter than the others. I think
leaving it to be either 1/8th (50% duration) or 1/16th (25% duration) is
sufficient.

I am not fully pleased with the ux of the dropdown menu for these two
new features, in general the design should stay the same but I am unsure
of why it is quite so laggy.